[Setting] #34 - 로그인, 온보딩 모듈 통합
