+++ Projects/Login/LoginData/Sources/Onboarding/DTO/CheckNicknameDuplicate.swift
@@ -1,0 +1,19 @@
+//
+//  CheckNicknameDuplicate.swift
+//  LoginData
+//
+//  Created by 김진웅 on 11/10/25.
+//  Copyright © 2025 Livith. All rights reserved.
+//
+
+// MARK: - 34. 닉네임 중복 확인
+
+import Foundation
+
+import LivithNetwork
+
+public extension DTO.Response {
+    struct CheckNicknameDuplicate: Decodable {
+        let available: Bool
+    }
+}
