+++ Projects/Login/LoginDomain/Sources/Onboarding/Repository/OnboardingRepository.swift
@@ -1,0 +1,14 @@
+//
+//  OnboardingRepository.swift
+//  LoginDomain
+//
+//  Created by 김진웅 on 11/10/25.
+//  Copyright © 2025 Livith. All rights reserved.
+//
+
+import Foundation
+
+public protocol OnboardingRepository {
+    func checkNicknameDuplicate(_ nickname: String) async throws(OnboardingError) -> Bool
+    func signup(nickname: String) async throws(OnboardingError)
+}
