+++ Projects/Login/LoginFeature/Sources/Onboarding/Model/NicknameValidationState.swift
@@ -1,0 +1,18 @@
+//
+//  NicknameValidationState.swift
+//  LoginFeature
+//
+//  Created by 김진웅 on 11/10/25.
+//  Copyright © 2025 Livith. All rights reserved.
+//
+
+import Foundation
+
+enum NicknameValidationState {
+    case idle
+    case valid
+    case invalid
+    case checking
+    case available
+    case duplicate
+}
