--- Projects/Onboarding/OnboardingData/Sources/DTO/CheckNicknameDuplicate.swift
@@ -1,19 +1,0 @@
-//
-//  CheckNicknameDuplicate.swift
-//  OnboardingData
-//
-//  Created by 김진웅 on 11/10/25.
-//  Copyright © 2025 Livith. All rights reserved.
-//
-
-// MARK: - 34. 닉네임 중복 확인
-
-import Foundation
-
-import LivithNetwork
-
-public extension DTO.Response {
-    struct CheckNicknameDuplicate: Decodable {
-        let available: Bool
-    }
-}
