--- Projects/Onboarding/OnboardingDomain/Sources/UseCase/Protocol/OnboardingUseCase.swift
@@ -1,15 +1,0 @@
-//
-//  OnboardingUseCase.swift
-//  OnboardingDomain
-//
-//  Created by 김진웅 on 11/10/25.
-//  Copyright © 2025 Livith. All rights reserved.
-//
-
-import Foundation
-
-public protocol OnboardingUseCase {
-    func validateNicknameFormat(_ nickname: String) throws(OnboardingError)
-    func checkNicknameDuplicate(_ nickname: String) async throws(OnboardingError)
-    func signup(nickname: String) async throws(OnboardingError)
-}
