--- Projects/Onboarding/OnboardingFeature/Sources/Route/OnboardingRoute.swift
@@ -1,19 +1,0 @@
-//
-//  OnboardingRoute.swift
-//  OnboardingFeature
-//
-//  Created by 김진웅 on 11/8/25.
-//  Copyright © 2025 Livith. All rights reserved.
-//
-
-import Foundation
-
-import Routing
-
-enum OnboardingRoute: String, Routable {
-    case terms
-    case nickname
-    case signupFailed
-    
-    var id: String { rawValue }
-}
