+++ Tuist/ProjectDescriptionHelpers/Module/Module+TargetID.swift
@@ -14,7 +14,6 @@ public enum TargetID {
     case core(CoreModule)
     case login(LoginModule)
     case dsKit
-    case onboarding(OnboardingModule)
     case search(SearchModule)
     case user(UserModule)
 
@@ -24,7 +23,6 @@ public enum TargetID {
         case .core(let module): return module.rawValue
         case .dsKit: return "DSKit"
         case .login(let module): return module.rawValue
-        case .onboarding(let module): return module.rawValue
         case .search(let module): return module.rawValue
         case .user(let module): return module.rawValue
         }
@@ -48,8 +46,6 @@ public enum TargetID {
         case .dsKit:
             return ["Sources/**"]
         case .login(let module):
-            return ["\(module.rawValue)/Sources/**"]
-        case .onboarding(let module):
             return ["\(module.rawValue)/Sources/**"]
         case .search(let module):
             return ["\(module.rawValue)/Sources/**"]
