+++ Tuist/ProjectDescriptionHelpers/Module/TargetDependency+Extension.swift
@@ -29,11 +29,7 @@ extension TargetDependency {
     public static func external(_ dependency: ExternalDependency) -> TargetDependency {
         return .external(name: dependency.rawValue)
     }
-
-    public static func onboarding(_ module: OnboardingModule) -> TargetDependency {
-        return .project(target: module.rawValue, path: ProjectID.onboarding.path)
-    }
-
+    
     public static func user(_ module: UserModule) -> TargetDependency {
         return .project(target: module.rawValue, path: ProjectID.user.path)
     }
